CityListPresenter.kt deleted and function moved to ViewModel
